Add util functions
Adds a function, "removeIdenticalLetters". It is a pure function that
takes a string and returns a new string where consecutive identical
letters are removed if they are more than a specified limit which
defaults to 4. I.e removeIdenticalLetters('aaaa') returns 'aaa'.

Some benchmarks were collected to compare two implementations of the
function, one using a reduce loop and one using a for-loop.
The for-loop implementation were aproximately 2.70x faster when
supplied with strings that were 150 000 characters long and it was
1.91x faster when supplied with string that were 20 000 000 characters
long. Hence, the for-loop implementation is used in this commit.

Furthermore, adds a function "maximiumOddSum". It is a pure function
that takes a list of numbers and finds the greatest odd sum from two of
these numbers. I.e. maximumOddSum([1,2,3,4,5]) returns 9.

Some benchmarks are added for this function to compare perfomance curves
when supplied with very long lists.